refactor auth api
