+++ packages/common/src/authentication/api/hooks/useAuthApi.ts
@@ -1,0 +1,9 @@
+import { useOmSupplyApi } from '../../../api';
+import { getSdk } from '../operations.generated';
+
+export type AuthApi = ReturnType<typeof getSdk>;
+
+export const useAuthApi = (): AuthApi => {
+  const { client } = useOmSupplyApi();
+  return getSdk(client);
+};
