+++ packages/common/src/authentication/api/operations.graphql
@@ -1,0 +1,73 @@
+query authToken($username: String!, $password: String!) {
+  authToken(password: $password, username: $username) {
+    ... on AuthToken {
+      __typename
+      token
+    }
+    ... on AuthTokenError {
+      __typename
+      error {
+        ... on UserNameDoesNotExist {
+          __typename
+          description
+        }
+        ... on InvalidCredentials {
+          __typename
+          description
+        }
+        ... on DatabaseError {
+          __typename
+          description
+          fullError
+        }
+        ... on InternalError {
+          __typename
+          description
+          fullError
+        }
+        description
+      }
+    }
+  }
+}
+
+query refreshToken {
+  refreshToken {
+    ... on RefreshToken {
+      __typename
+      token
+    }
+    ... on RefreshTokenError {
+      __typename
+      error {
+        description
+        ... on DatabaseError {
+          __typename
+          description
+          fullError
+        }
+        ... on TokenExpired {
+          __typename
+          description
+        }
+        ... on NotARefreshToken {
+          __typename
+          description
+        }
+        ... on NoRefreshTokenProvided {
+          __typename
+          description
+        }
+        ... on InvalidToken {
+          __typename
+          description
+        }
+        ... on InternalError {
+          __typename
+          description
+          fullError
+        }
+      }
+    }
+  }
+}
