+++ packages/common/src/authentication/index.ts
@@ -1,1 +1,2 @@
 export * from './AuthContext';
+export * from './api';
