+++ packages/common/src/operations.graphql
@@ -236,80 +236,6 @@ query stores($first: Int, $offset: Int, $filter: StoreFilterInput) {
   }
 }
 
-query authToken($username: String!, $password: String!) {
-  authToken(password: $password, username: $username) {
-    ... on AuthToken {
-      __typename
-      token
-    }
-    ... on AuthTokenError {
-      __typename
-      error {
-        ... on UserNameDoesNotExist {
-          __typename
-          description
-        }
-        ... on InvalidCredentials {
-          __typename
-          description
-        }
-        ... on DatabaseError {
-          __typename
-          description
-          fullError
-        }
-        ... on InternalError {
-          __typename
-          description
-          fullError
-        }
-        description
-      }
-    }
-  }
-}
-
-query refreshToken {
-  refreshToken {
-    ... on RefreshToken {
-      __typename
-      token
-    }
-    ... on RefreshTokenError {
-      __typename
-      error {
-        description
-        ... on DatabaseError {
-          __typename
-          description
-          fullError
-        }
-        ... on TokenExpired {
-          __typename
-          description
-        }
-        ... on NotARefreshToken {
-          __typename
-          description
-        }
-        ... on NoRefreshTokenProvided {
-          __typename
-          description
-        }
-        ... on InvalidToken {
-          __typename
-          description
-        }
-        ... on InternalError {
-          __typename
-          description
-          fullError
-        }
-      }
-    }
-  }
-}
-
 query masterLists(
   $first: Int
   $offset: Int
